+++ src/commands/general/msfsdiscord.ts
@@ -1,4 +1,4 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 const MSFS_DISCORD_LINK = 'https://discord.gg/msfs';
@@ -7,5 +7,5 @@ export const msfsdisc: CommandDefinition = {
     name: ['msfsdisc', 'fsdisc', 'msfsdiscord'],
     description: 'Provides link to Microsoft Flight Simulator discord server',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send(MSFS_DISCORD_LINK),
+    executor: (msg) => replyWithMsg(msg, MSFS_DISCORD_LINK),
 };