+++ src/commands/general/website.ts
@@ -1,4 +1,4 @@
-import { CommandDefinition } from '../../lib/command';
+import { CommandDefinition, replyWithMsg } from '../../lib/command';
 import { CommandCategory } from '../../constants';
 
 const WEBSITE_URL = 'https://flybywiresim.com/';
@@ -7,5 +7,5 @@ export const website: CommandDefinition = {
     name: ['website', 'web'],
     description: 'Links to the FlyByWire Simulations website.',
     category: CommandCategory.GENERAL,
-    executor: (msg) => msg.channel.send(WEBSITE_URL),
+    executor: (msg) => replyWithMsg(msg, WEBSITE_URL),
 };