+++ src/commands/moderation/temporaryCommandEdit.ts
@@ -221,7 +221,7 @@ export const temporarycommandedit: CommandDefinition = {
         const subCommands = ['add', 'image', 'delete', 'info'];
         const conn = await getConn();
         if (!conn) {
-            return msg.channel.send({ embeds: [noConnEmbed] });
+            return msg.reply({ embeds: [noConnEmbed] });
         }
 
         const modLogsChannel = msg.guild.channels.resolve(Channels.MOD_LOGS) as TextChannel | null;
@@ -229,7 +229,7 @@ export const temporarycommandedit: CommandDefinition = {
         const evokedCommand = msg.content.split(/\s+/)[0];
         const args = msg.content.replace(evokedCommand, '').trim();
         if (!args || args === 'help') {
-            return msg.channel.send({ embeds: [helpEmbed(evokedCommand)] });
+            return msg.reply({ embeds: [helpEmbed(evokedCommand)] });
         }
 
         let [subCommand] = args.split(/\s+/);
@@ -243,19 +243,19 @@ export const temporarycommandedit: CommandDefinition = {
             const regexCheck = /^["]?\.?(?<command>[\w-]+)["]?\s["]?(?<severity>info|warning|error)["]?\s"(?<title>[^"]*|^[^"]*$)"\s"(?<content>[^"]*|^[^"]*$)"\s*$/;
             const regexMatches = subArgs.match(regexCheck);
             if (regexMatches === null || !regexMatches.groups.command || !regexMatches.groups.severity) {
-                return msg.channel.send({ embeds: [missingInfoEmbed('Add', `You need to provide the expected format to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
+                return msg.reply({ embeds: [missingInfoEmbed('Add', `You need to provide the expected format to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
             }
             if (regexMatches.groups.title === '') {
-                return msg.channel.send({ embeds: [missingInfoEmbed('Add', `You need to provide a non-empty title to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
+                return msg.reply({ embeds: [missingInfoEmbed('Add', `You need to provide a non-empty title to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
             }
             if (regexMatches.groups.content === '') {
-                return msg.channel.send({ embeds: [missingInfoEmbed('Add', `You need to provide a non-empty content to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
+                return msg.reply({ embeds: [missingInfoEmbed('Add', `You need to provide a non-empty content to create a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
             }
             const { command, content, title, severity } = regexMatches.groups;
             const searchResult = await TemporaryCommand.find({ command });
 
             if (searchResult.length !== 0) {
-                return msg.channel.send({ embeds: [alreadyExistsEmbed('Add', command)] });
+                return msg.reply({ embeds: [alreadyExistsEmbed('Add', command)] });
             }
 
             const moderator = msg.author;
@@ -274,7 +274,7 @@ export const temporarycommandedit: CommandDefinition = {
             try {
                 await temporaryCommand.save();
             } catch {
-                return msg.channel.send({ embeds: [failedEmbed('Add', command)] });
+                return msg.reply({ embeds: [failedEmbed('Add', command)] });
             }
 
             const { imageUrl } = temporaryCommand;
@@ -298,7 +298,7 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Add', 'Mod Log')] });
+                await msg.reply({ embeds: [noChannelEmbed('Add', 'Mod Log')] });
             }
 
             try {
@@ -317,23 +317,23 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Add', 'Support')] });
+                await msg.reply({ embeds: [noChannelEmbed('Add', 'Support')] });
             }
 
-            return msg.channel.send({ embeds: [successEmbed('Add', command)] });
+            return msg.reply({ embeds: [successEmbed('Add', command)] });
         }
 
         if (subCommand === 'image') {
             const regexCheck = /^["]?\.?(?<command>[\w-]+)["]?\s["]?(?<imageUrl>https?:\/\/[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*))["]?\s*$/;
             const regexMatches = subArgs.match(regexCheck);
             if (regexMatches === null || !regexMatches.groups.command || !regexMatches.groups.imageUrl) {
-                return msg.channel.send({ embeds: [missingInfoEmbed('Configure image URL', `You need to provide the expected format to set the image for a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
+                return msg.reply({ embeds: [missingInfoEmbed('Configure image URL', `You need to provide the expected format to set the image for a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
             }
 
             const { command, imageUrl } = regexMatches.groups;
             const temporaryCommandSearchResult = await TemporaryCommand.find({ command });
             if (temporaryCommandSearchResult.length === 0) {
-                return msg.channel.send({ embeds: [notFoundEmbed('Configure image URL', command)] });
+                return msg.reply({ embeds: [notFoundEmbed('Configure image URL', command)] });
             }
             const [temporaryCommand] = temporaryCommandSearchResult;
 
@@ -341,7 +341,7 @@ export const temporarycommandedit: CommandDefinition = {
             try {
                 temporaryCommand.save();
             } catch {
-                return msg.channel.send({ embeds: [failedEmbed('Configure image URL', command)] });
+                return msg.reply({ embeds: [failedEmbed('Configure image URL', command)] });
             }
 
             const { date, moderator, severity, title, content, lastUsed } = temporaryCommand;
@@ -365,7 +365,7 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Configure image URL', 'Mod Log')] });
+                await msg.reply({ embeds: [noChannelEmbed('Configure image URL', 'Mod Log')] });
             }
 
             try {
@@ -384,28 +384,28 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Configure image URL', 'Support')] });
+                await msg.reply({ embeds: [noChannelEmbed('Configure image URL', 'Support')] });
             }
 
-            return msg.channel.send({ embeds: [successEmbed('Configure image URL', command)] });
+            return msg.reply({ embeds: [successEmbed('Configure image URL', command)] });
         }
 
         const regexCheck = /^["]?\.?(?<command>[\w-]+)?["]?.*$/;
         const regexMatches = subArgs.match(regexCheck);
         if (!regexMatches || !regexMatches.groups || !regexMatches.groups.command) {
             const subCommandText = subCommand === 'info' ? 'show the info of' : subCommand;
-            return msg.channel.send({ embeds: [missingInfoEmbed(subCommand, `You need to provide the expected format to ${subCommandText} a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
+            return msg.reply({ embeds: [missingInfoEmbed(subCommand, `You need to provide the expected format to ${subCommandText} a temporary command. Check \`${evokedCommand} help\` for more details.`)] });
         }
         const { command } = regexMatches.groups;
         if (subCommand === 'info') {
             const searchResult = await TemporaryCommand.find({ command });
 
             if (searchResult.length === 0) {
-                return msg.channel.send({ embeds: [notFoundEmbed('Info', command)] });
+                return msg.reply({ embeds: [notFoundEmbed('Info', command)] });
             }
             const { author } = msg;
             const { moderator, content, date, title, severity, imageUrl, lastUsed } = searchResult[0];
-            return msg.channel.send({
+            return msg.reply({
                 embeds: [infoEmbed(
                     temporaryCommandEmbedField(
                         timestampText(date),
@@ -428,16 +428,16 @@ export const temporarycommandedit: CommandDefinition = {
             try {
                 temporaryCommands = await TemporaryCommand.find({ command });
                 if (!temporaryCommands || temporaryCommands.length !== 1) {
-                    return msg.channel.send({ embeds: [notFoundEmbed('Delete', command)] });
+                    return msg.reply({ embeds: [notFoundEmbed('Delete', command)] });
                 }
             } catch {
-                return msg.channel.send({ embeds: [notFoundEmbed('Delete', command)] });
+                return msg.reply({ embeds: [notFoundEmbed('Delete', command)] });
             }
 
             try {
                 temporaryCommands[0].delete();
             } catch {
-                return msg.channel.send({ embeds: [failedEmbed('Delete', command)] });
+                return msg.reply({ embeds: [failedEmbed('Delete', command)] });
             }
 
             const { author } = msg;
@@ -462,7 +462,7 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Delete', 'Mod Log')] });
+                await msg.reply({ embeds: [noChannelEmbed('Delete', 'Mod Log')] });
             }
 
             try {
@@ -481,12 +481,12 @@ export const temporarycommandedit: CommandDefinition = {
                     )],
                 });
             } catch {
-                await msg.channel.send({ embeds: [noChannelEmbed('Delete', 'Support')] });
+                await msg.reply({ embeds: [noChannelEmbed('Delete', 'Support')] });
             }
 
-            return msg.channel.send({ embeds: [successEmbed('Delete', command)] });
+            return msg.reply({ embeds: [successEmbed('Delete', command)] });
         }
 
-        return msg.channel.send({ embeds: [helpEmbed(evokedCommand)] });
+        return msg.reply({ embeds: [helpEmbed(evokedCommand)] });
     },
 };