+++ src/commands/moderation/warn/listWarnings.ts
@@ -21,7 +21,7 @@ export const listWarnings: CommandDefinition = {
         const conn = await getConn();
 
         if (!conn) {
-            await msg.channel.send({ embeds: [noConnEmbed] });
+            await msg.reply({ embeds: [noConnEmbed] });
             return;
         }