+++ src/commands/support/utf-8.ts
@@ -7,7 +7,7 @@ const UTF8_HELP_URL = `${process.env.IMAGE_BASE_URL}support/utf-8.PNG`;
 const utf8Embed = makeEmbed({
     title: 'FlyByWire A32NX | UTF-8',
     description: makeLines([
-        'Some users experience problems with various system in the A32NX. These are caused by an issue within MSFS which requires the use of the UTF8 Region setting in Windows.',
+        'Some users experience problems with various system in the A32NX. These are caused by an issue within MSFS which requires the use of the UTF-8 Region setting in Windows.',
         '',
         'To enable UTF-8 support follow the steps below:',
         '',
@@ -25,7 +25,7 @@ const utf8Embed = makeEmbed({
 });
 
 export const utf8: MessageCommandDefinition = {
-    name: 'utf8',
+    name: 'UTF-8',
     description: 'Provides a link to resolve UTF-8 issues',
     category: CommandCategory.SUPPORT,
     genericEmbed: utf8Embed,